Fixes for telemetry
It could be my implementation, but I had to make these changes to
get telemetry working.  Service type for me is metering.  If that
is not always the case, maybe there should be the ability to
supply multiple types or override through some mapping.  Also,
the v2.0 url did not work for me.

Change-Id: Ia664b5d6e142a1c9f05134ccb5222451de9319a1