Change the capabilities to capability
I was looking at the meter resource when I noticed the
capabilties were not working properly for list.  After I saw
the output I was getting from the ceilometer I have to mess with,
I thought it should be changed around a bit:

* The resources_key does not appear in the output of get, instead
  'api' appears.
* The Resource.list method expects an array but the capabilities
  API delivers a dict.  So, I just overrode list.
* From the user perspective, I think it is more consistent to
  have list return a list of capability instead of one
  capabilities.

It would be nice if the Ceilometer folks did something different
here next time around.

Change-Id: I27c42882de6bb36f83bdc36193322035d0f8114f