Add clouds-public.yaml
Put vendor config outside of the code in clouds-public.yaml. Fall back to
vendors.py if clouds-public.yaml not found.

The search follows the same rules as clouds.yaml, the file is the same
format except the top-level key is 'public-clouds'.  Typically only auth_url
and region_name are populated.