Map CloudConfig attributes to CloudConfig.config
Treat the CloudConfig object as if it has the config attributes directly.
And add some simple tests.

This makes it easier to replace an argparse.Namespace() object with a
CloudConfig object.  It also might make initialization of some of the
default attributes unnecessary.  An example of this usage is in
https://review.openstack.org/#/c/129795/1/openstackclient/shell.py

Change-Id: I00ced540cf94742e8cb738f8f0767445ffeb4bfe