High level interface
High level interface using the connection classes to access the
various services.  Only identity projects and list_flavors in
compute are implemented:

    projects = conn.identity.list_projects()

The service filter has a valid_versions added to it to map the
module name to the version component of the endpoint URL.  For
example, the v2.0 endpoint path is supported by the v2 SDK
network module:

https://region-a.geo-1.network.hpcloudsvc.com/v2.0/

is supported by the SDK module openstack.network.v2.

Implements: blueprint highlevel-interface

Change-Id: Ie52986a9da84c8f3395757c08de7a01be30e0f2a