Implement iterator paging
Instead of manually passing in limit and marker, automatically set the
marker to the end of the received data and continue requesting data
until none is returned. This is a naive implementation as it doesn't
attempt to know when it has received all of the data without making one
request past the end, as pagination is implemented in several different
ways across the services. Optimizations to reduce that extra call should
probably be implemented per-service due to those differences.

Change-Id: I63a8eb7836c1a614e4bc2ee98fa93f239b343cbb