Initial "Getting Started" guide
This guide lays out the necessary parts for an existing OpenStack
developer to get started. It is admittedly not great for those coming
from the outside, which I am absolutely interested in fixing, but I want
to make sure those who are immediately interested in contributing from
the service side have what they need to get going.

Change-Id: I617b683ab2cbf7856bd16ea8e812e5311c8064b0