Sample thin interface
Here is an example of a thin interface for the SDK.  We aren't
making any resources here, it is all dictionaries.  Assuming we
have a session:

    from openstack.network.v2 import thin
    requests = thin.Thin()
    for dictobj in requests.list_networks(session):
        print(dictobj['id'])

Obviously, the thin layer isn't doing much, but it does provide
a simple interface to access the methods.  There may be cases
when it would need to do more.

If we go with this approach, we'd just neeed to be sure that if
a resource needed special handling for an operation, we override
the class method rather than the object method.

Change-Id: I2b41ad64348802f6461abfd39a157e808686bb4d