Build Resource from either existing object or id
Allow Resource objects to be built from either an existing object, or be
created from scratch with the given id. This is useful at the Proxy
level when creating interfaces that must work with a Resource in the
end, but will want to be able to either operate on string names or the
Resources returned from other APIs.

An example is that you may want to create an object in the object store
given a previously obtained container object and a string name for the
object. You may also have constructed the Resource for the object ahead
of time and may want to pass that instead of a string name.

Change-Id: I84c2101024849ceb78de677607f6a64bd8134ede