Add coverage-package-name to tox.ini for coverage
Recent pbrs support a coverage-package-name argument which allows you to
explicitly specify the package name to test coverage of. By default it
takes the project name (python-openstacksdk) and splits the "python-"
prefix off, leaving it to check coverage of the nonexistent
"openstacksdk" package. After this, "tox -e cover html" works and drops
the results in a cover/ folder under root.

Change-Id: I55e3fb274158aa37938329f2153c5c9597e64915