Complete the Resource class documentation
Document the CRUD methods with explanations of arguments, return value,
and exceptions that could be raised. This also introduces a change to
the autoclass setting so that it orders the members by how they are
written in the source, instead of alphabetical order.

A second change is going to be submitted that does some reordering
of methods so that more common methods will be near the top and the more
internal (but still public) methods near the bottom.

Change-Id: I94facb4d6c4e8d446e8e696afb75f1a3268d3722