Prepare for documentation of Resources
Add a section for where Resource classes will show up in the docs and
enter two skeleton pages for how this will look. Later changes will
cover the actual member-level documentation of the Resource classes for
each service, along with their addition to the TOC.

This additionally sets autodoc_member_order globally to "bysource",
meaning it'll order the rendered member documentation in the order it
was written. The default of alphabetical order by what shows up in
__dir__ is not as helpful or as readable as how we've organized the code.

Change-Id: Ifcd7cd81ba55ecb429c50680a926d4f8a44ca9a3