Don't attempt to __get__ prop without instance
Doing something like Container.name raises an AttributeError because
we're trying to get the `_attrs` attribute of an instance that is of
NoneType. In general that's perhaps an odd behavior, but a primary issue
is that it hampers autodoc from being able to access member names on a
class, since it has to be able to import classes and access each member.
Since the members we're documenting are resource.props, which operate on
instances, we should bail out early if we're not actually given an
instance.

Change-Id: I45be106f88084325618df64e871b75accb36bf9f