Add image upload support
There are at least 2 known mechanisms in the wild for uploading images
to public glance endpoints that are completely different. Based on which
version of the API we're running switch between them seamlessly. Note
that the wait parameter isn't going to do anything for v1 because it's a
sync call - maybe we should throw an error if someone tries to leave if
set to false?

Change-Id: Ifccf8aaa6cd7793a1b827914e7da4b2b82248544