Debug log any time we re-raise an exception
We're masking exceptions on purpose, which is evil of course - but the
reason is that they wind up being part of the API - and the
python-*client-ness of a particulate exception is largely an
implementation detail that might change. Eventually we want to have
better exceptions across the board - but make sure that we don't lose
the ability to debug the exceptions when they happen.

Change-Id: I8a972e6608e33088234d67ea24886c782a80e3b4