Adjust some parameters and return types
This change removes a TODO from get_account_metadata which was
previously taking a Container when it wasn't necessary (and didn't make
a ton of sense). Most of the other changes are to document return types
which were previously unlisted, and in a few cases of requests which
don't return a body, their responses are turned to None. If the request
worked, you're going to get an unchanged object back, so no need to be
returning the same thing back at this level.

Change-Id: Ic547e6912e14dfe622533550f1432415070566c1