Reorganize existing documentation files
This change creates two directories of content, based on who they're
relevant to: users and contributors. Files which were in the top-level
folder were moved, as well as some within folders in that top-level.
This is one change of a multi-step process, so it only deals with
renamings in order to make further changes cleaner.

Change-Id: I3c9bb7da4f283e2696e7bcea84824c778bbcecfb
Partial-Bug: 1416553