Support Resource as a type for properties
Currently we only support built-in types, but given that we send and
receive IDs which correspond to Resource types we have, this adds a case
that knows how to work with Resource and its subclasses. This allows us
to work with Resource objects in the most simple case of being returned
an ID, but also allows us to build more complex Resources, such as the
Limits resource in Compute v2.

Change-Id: Ib3c0634e4f750f44efe2a74eb00cc4c1ad8d2a5d