Add logging functionality to openstack.utils
This change moves the ability to enable logging out of the
examples.common module and into the openstack.utils since it is
generally useful.

This is intended to give users a short-term solution to receive
debugging messages. If they want to always have our messages logged in
their application they should create their own logging handler(s) with
their own format, and that will cause our messages to be sent to their
handlers as they wish.
https://bugs.launchpad.net/python-openstacksdk/+bug/1418307 is open to
track the documentation aspect of this.

Change-Id: Ic8768d529d0389b9b04bd4ea31453a82f30faf33