Provide a better default user-agent string
This change provides a default user-agent with more information and
adjusts the handling of user-agent arguments to fall in line with RFC
7231 Section 5.5.3. We now specify the package name and version for the
SDK as well as the Python implementation and version as the default.

From there, a user can specify their application name and version but it
does not replace the overall user-agent. As the RFC states,
product/version pairs are listed in decreasing order, so anything the
user specifies is prepended to whatever string is currently set for that
Transport instance.

If headers are passed in to a Transport.request method which carry a
`User-Agent`, they are used as-is unless the caller also specified a
`user_agent` parameter.

This change simplifies the tests slightly without reducing any coverage.

Change-Id: I80ca26fff3f2522b8232472676396abb86166f91