Allow region_name to be None
When using the shade library with ansible playbooks and classical
environment variables for authentication, it is possible for a
region_name value to be passed in with the value of None.

When the value is set to and preserved as None, the logic later
on in the method will fail to create a config key/value entry for
region_name which is required by code later on in the method.

Modified the region_name check to not only check for the presence
of a missing region_name value, but to check to see if that value
is set to None.  This allows the value to be reset preventing the
method from failing on the return call due to a missing key.

Change-Id: Id8a3edf53ac751f0c6ee4d71405a926ba90c0602