Set Resource.page limit to None
Setting the Resource.page limit to None, which is how it gets used by
default from Resource.list, allows it to be used directly with the same
ease as Resource.list. For APIs which return lists but don't support
pagination, this allows us to provide users the same iteration behavior
across the board.

For an example of where this comes into play, a GET on /types in the
volume service returns a list of types in a single response with
no pagination support. If we call Types.list(session) it will be an
infinite loop. With this change, Types.page(session) allows the Type
proxy to yield type objects out of that one page.

Change-Id: I242a9fd0cd6d21b891d91e2fe69006a1f0ec3a15