Add status_code to HttpException
While coding up Volume V2, while implementing the ability to get one
type at a time, e.g., GET /types/12345, it'll return a 404 if the type
doesn't exist. I'd like to convert that into None so the caller either
gets the type or they don't. In order to do that we have to be able to
continue passing on the status code from the exception raised inside of
transport.

This change not only exposes status_code, but it enables the ability to
get the exception message as an attribute on the base SDKException. That
was noticed while writing up tests.

Change-Id: Id0e067a77d9e33600c9f8f07e95b6130c0b1bd0a