Determine limit based on page size
Without a limit parameter being provided, Resource.list required making
one call past the end of data in order to know it was done -- when it
received an empty response body.

This change figures out the limit based on the count of items returned,
then uses that in subsequent requests. The bigger gain here is that we
now know how many items constitutes a full page, so we can stop making
requests for more pages once we've received a smaller page than the
known limit.

Change-Id: I5b7739f14a8934b6c7a53987a2aec0f251fb240b