Bypass type conversion when setting default prop
When a prop has a type, if the default value is ever put through
__set__, we'll try to convert that default value into type. For example,
as is currently happening in network.v2.router, the REST API returns
null (None) when no gateway information is set, but when it is set, it
returns a dictionary. Trying to put that default None value through to
set the value causes a TypeError.

If we're given the default value to set on a property, just
short-circuit and return out of __set__. We already have the default
and we know how to return the default value.

The reason we don't go through and set this default value into
instance._attrs is that it puts us through type conversion on the
__get__ end of it. Actually setting this default value is that true
internal attribute value will then raise a TypeError during the
attribute access.

Change-Id: I106c9c54f57596cb2e12e8bbb5cd380867be876b