Move metric capabilities into a single Resource
Previously, one call to /capabilities would result in potentially
several Capability objects being returned (through `list`),
and you had to check the value to know what type of capability
it was. This change brings that all together in one object and
sets the type of capabiltiy returned as a property on the object.

It now works like this:

>>> from openstack.metric.v1 import capabilities
>>> caps = capabilities.Capabilities().get(session)
>>> caps.aggregation_methods
['mean', 'max', 'avg']

Change-Id: Icee03330d42df8aeea11863eef7e4721d8311156