Add Resource.name property
We've been able to set name_attribute for a while now, but it's not
actually hooked up to anything. We've also been attempting to access it
through Resource.find and getting lucky that resources typically do have
a name property set. This change adds it so we can reliably access it
and provide the fallback that we've sort of been providing all along.
You can also set the name here, which sets the underlying attribute
pointed to by name_attribute.

Change-Id: I081816adebbebf26339b5b57e3f79fc7b10649f7