Create server requires *Ref names during POST
The compute service expects imageRef and flavorRef in the request body,
so combined with change 156101 to store values in resource.prop as
they're named, this change allows Server.create to work properly when
constructed from its props. Previously it was sending the values as
image and flavor.

Moving image to being an alias allows everything else to work as
expected.

Change-Id: I3705149d7f5396152c98e4e78fb7c68518db1efc