Prefer storing prop values as their name
When a resource.prop is named the same as its alias, the Resource
initializer inserts an entry with that alias name into its
_attrs dictionary, and then it attempts to call setattr using
that same name, which is also the name of the resource.prop. When this
happens, _attrs has a key for the `alias` and it then tries to add an
entry for the `name` of the prop. Since this is happening in a loop
around the _attrs dictionary, it causes the dictionary to increase in
size, which causes a RuntimeError. Here's an example that would cause
this to occur.

    attr = resource.prop("something", alias="attr")

If we prefer to always store the value under the `name`, in this case,
something, we won't ever attempt to mutate the dictionary while setting
props.

A slightly similar issue came up before, where a response body
contained the exact name of a prop which had a different name, so
the same mutation during __set__ was occurring.

Change-Id: Iee5c1f050c9bfcfc5afa8bbf77df7e1db02564d3
Related-Bug: 1414089