Use case insensitive dict for Resource attributes
Any resources which work with headers - mostly Container and Object in
object_store - are not currently working properly due to the case
sensitive nature of the attributes being in a traditional dictionary.
Any header props *set* on those resources which are then sent to the
server are ok, which is why we've been able to set and then receive our
own values,  but any that are received would have to be in the matching
case that the prop is setup as. This is not generally the case, as
setting metadata on containers or objects in Horizon or other control
panels uses capitalization, which ends up not matching up properly when
we convert the server values into a resource.

This change implements a CaseInsensitiveDict which keeps the original
user-set values around and then keeps a dictionary mapping a lowercased
version of the key to the user-set key. It implements the typical
MutableMapping operations.

Change-Id: Ic815b633bf0bf8201117740ca40ceda600856734