Allow Resource attributes to be updated
With Resource being a MutableMapping, it inherits an `update` method
like any mapping does. In the Connection layer, this would be really
useful when we want to be able to update properties on a Resource
based on dictionaries or named arguments users are calling proxy
methods with. However, since we override `update` to do our own idea
of updating, we lose that.

A specific example of where this is useful is that
object_store.set_*_metadata would work great if it could take named
arguments and set them on a resource, and then call `update`. The
following pseudocode explains:

    def set_metadata(container, **metadata):
        container.update_attrs(metadata)
        ...

    set_metadata(c, read_ACL="foo", write_ACL="bar")

This would result in those resource.props being set on the `c`, whereas
today we would have to look into the **metadata keys and values and send
them through setattr.

Change-Id: I1f7d71e7874ad39c353e299e3d4aac11c5e229e8