Mutate Resource via mapping instead of dict
resource.prop was doing all of its mutation of the parent Resource via
instance._attrs, i.e., directly modifying the attributes and working
around the MutableMapping magic methods where we've handled things like
the dirty list, which we then look at when doing calls like
Resource.update. Because we're bypassing that, we're not properly
sending the correct values in update calls for attributes that were
modified via their prop.

If we make resource.prop do its get/set/delete on the Resource instance
and let the Resource instance work with the underlying dictionary, the
dirty list is kept accurate regardless of how attributes were modified,
whether it's through creation of the object, prop update, or mapping
update.

Change-Id: I977ce898a4a7b8718d7d6a29470d51ab8a5dd674