Support listing non-paginated response bodies
We previously tried to discover whether or not a resource supported
pagination by making extra calls, and then another attempt was made to
separate the idea of paginated and unpaginated by making separate calls
and modifying Resource.page. We can instead pass in a pagination
parameter to list, defaulting to True, which remains the one call to
make in order to be yielded Resource objects.

Change-Id: Ie97ef5b43fcabef145ab2758082ae9f2f400a0ab
Related-Bug: 1410980