Serialize Resource types before sending to server
Now that we can have Resource instances as the type of resource.prop, we
need to be able to serialize them before sending them to the server. We
can receive them just fine as responses, but need to get their IDs out
so that the attributes we send in create and update requests work
properly.

This change adds a method on the Resource class that takes the
attributes we're about to create or update and checks if any of them are
Resources. If they are, it replaces that item in the dictionary with its
ID.

Change-Id: I6501f167c31b0df1e5ed7c85d2d7a646e537f832