Add FlavorDetail for extra information
The current Flavor can only receive the basics returned by a GET
/flavors, which is just name, id, and links. A call to /flavors/detail
gives a number of added details, some of which were already in the
Flavor resource but not set. This change follows with how a few other
basic/detail requests were done, and exposes it through the proxy with
details coming by default (since they're likely expected, and most
useful).

This change depends on list handling nonpaginated bodies.t

Change-Id: I7bebbdc2aa86c0300b86a8b8daa34888121dbcff