Add ImageDetail for extra information
The current Image can only receive the basics returned by a GET
/images, which is just name, id, and links. A call to /images/detail
gives a number of added details, some of which were already in the
Image resource but not set. This change follows with how a few other
basic/detail requests were done, and exposes it through the proxy with
details coming by default (since they're likely expected, and most
useful).

Change-Id: Iadc98cd0f679777b1c6e14711ad00787142221e2