Send empty dict when no headers on create/update
When no headers are set on a Container.create, we end up attempting to
send `None` through to the various lower levels, and they depend on
being able to set things like the token on the headers that get sent
through. If `attrs` sent into both Container.create and Container.update
don't actually contain any headers, put an empty dictionary in place of
them.

Change-Id: Iea050976d86a12d7b78245147d3560b188c8a584