Allow keystone validation bypass for noauth usage
To enable the eventual use of the shade library for communication
with ironic in situations where authentication is set to noauth, it
is necessary to not attempt to validate a user's credentials.

Added a check to disable authentication validation when the
auth_plugin is set to '' or 'None' or None.

Change-Id: I3807b4724ce5e204b5857c1dbf5325f0e3f4a78d