Adjust paginate argument usage
When we changed the default from paginating list to not paginating list,
a few other uses of the argument were not updated for that swap.
However, since the default turned to False, nothing actually broke in
IRL quick tests, but anything that required pagination would not have
been returning the full amount of data. Additionally, during that
change, the argument name went from paginate to paginated and wasn't
fully changed to match up.

Change-Id: Id69f27ed30beb55eff96700f05075852525f936f