Align cert, key, cacert and verify with requests
The semantics of these values are specific, and an interface we'd
carried over from the old ansible modules really just makes things more
confusing.

Requests expects that verify can either be True/False or can be the path
to a cacert bundle. That makes it very hard to do input parameter
validation in ansible. So - split that into verify which takes a boolean
and a path to a cacert bundle.

Requests also expects that cert can either be a path to a cert/key
bundle, or a tuple of two paths. I'm not really sure why it thinks
that's clever, but I don't.

Change-Id: Ie96e7afe3ef056499477b746d9262975081206d7