Add base for Proxy classes to inherit from
While exploring adding a parameter that each Proxy class would need, I
was reminded that we have to manually ensure they all take the same
parameters. Ideally we should just have them all inherit from one base,
and then when we load them, ensure they're all subclasses of that base.

This doesn't add any new tests because 1) it would require a ton of
mocking to the other things which are implicitly in the way (since this
is called from _open, which is called during __init__), and  that
doesn't seem worth it because 2) it's already in a code path that gets
tested by many other Connection related tests.

Change-Id: I953bd78f05e8d72a53d262b16c349d5d69da2fbf