Flesh out api version defaults
openstackclient needs a bit richer support for api version defaults.
Namely, it knows what defaults it wants to have - but we need to do
defaults processing in os-client-config to get sequencing correct. So
provide an API call to set new defaults that can be used before config
processing. Also, flesh out the dict of known default values with good
defaults to match osc behavior, and add the known v1 default of HP to
the vendors.py values.

Change-Id: I45e2550af58aee616ca168d20a557077beeab007