Improve image.v2.tag
While looking into create_by_id usage, I came across image.v2.tag as
something that needed to be adjusted. It was a bit cumbersome to use and
is doesn't benefit from a lot of the usual resource things since it
it just needs to join a string to the base_path and doesn't send or
receive anything else.

    t = tag.Tag({"image": image_resource_or_id})
    t.create(conn.session, "new tag")
    t.delete(conn.session, "old tag")

Building a proxy will be trivial for this. Viewing tags isn't offered by
this resource, but you can see them via image.Image.tags

Change-Id: I2bd549c55a8c21ac292752509122896a6fd51fe8