Proxy get method
This change introduces a common get API that can be applied to every
proxy method that revolves around get a single resource. In particular,
the change applies that base method to getting a single server.

The method takes either a resource or an ID and calls `get` on the
corresponding resource, after obtaining the canonical ID for that value
out of Resource.get_id. The ID is passed to the _check_resource
decorator so we're only making the get call on the appropriate type of
resource, if given. When given a string ID, we have to try it.

Change-Id: I2651d7e01b0067865d623dcda3b9b0f6bd43101b