Add new _verify to proxy base tests
This is a temporary addition to the tests which will allow us to verify
the new form of API calls. This is being done in lieu of adding logic
and extra arguments to the tests, which will unnecessarily complicate
them.

This was originally a part of the common_delete_method change, but is
needed by other later changes, so I want it out of delete so we don't
start a huge dependency chain on the delete change.

Change-Id: I27f9cfd2188acd31cee120d0347504b441fa61e9