Proxy delete method
This change introduces a common delete API that can be applied to every
proxy method that revolves around deleting a resource. In particular,
the change applies that base method to deleting a server.

The method takes either a resource or an ID and calls delete on the
corresponding resource, after obtaining the canonical ID for that value
out of Resource.get_id. This value is checked before making the request
to see if we can determine it's of the wrong resource type by the
_check_resource decorator.

Additionally, in the case of a 404, by default this is ignored, but if
ignore_missing is set to False, we handle a 404 and raise
exceptions.ResourceNotFound as a more descriptive exception.

Change-Id: I755c57e9bd9cec597a5bcf84c527e7bd6d710fb3