Add _check_resource to BaseProxy
All of the base methods on BaseProxy need to check their arguments in
order to prevent requests from being made on the wrong types, preventing
potentially confusing error messages.

This decorator covers two cases:

1. When strict=False, the `actual` parameter is checked to see if it's
   the `expected` type IFF `actual` is an instance of Resource. When
   the type does not match, ValueError is raised. If the type is
   something else, such as an id, it is passed on to the underlying
   method when it is able to additionally operate on an id, such as on a
   get or delete call.
2. When strict=True, the `actual` parameter must be an instance of
   Resource or None, otherwise it raises ValueError. This is necessary for
   methods such as create which can't operate solely on an id but
   require additional information.

In shorter terms: 1) if it's a resource we check it, 2) it has to be
a resource or nothing.

As the majority of operations can be carried out with just an ID, False
is the default strict mode.

Change-Id: I75d398fda55e80db5877bb7d55f7ff3d5dbd88ad