Proxy create method
This change introduces a common create API that can be applied to every
proxy method that revolves around creating a resource. In particular,
the change applies that base method to creating a server.

The method takes the attributes to create a resource and then calls the
create method on it.

As with the other similar changes to add base proxy methods, the
duplicate verification test will be removed once this API is applied to
the rest of the service proxies.

Change-Id: Ie718af8e81ec942c6d1dfa0f0b05becec88849d8