Proxy update method
This change introduces a common update API that can be applied to every
proxy method that revolves around updating a single resource. In
particular, the change applies that base method to updating a single
server.

The method takes either a resource or an ID and calls update on the
corresponding resource.

As with other similar changes, it temporarily introduces a second verify
test method. Once this is applied everywhere, it'll be removed.

Change-Id: I2ed6ada828034afe1f79d933cedb942c3ab57c2c