The first functional test
Just implementing a single functional test to start things off.

This test introduces a new test dependency on os-client-config.
os-client-config is used to configure the SDK to run its functional test
on a particular cloud.

To run the functional tests do a

    tox -e functional

Unit tests continue to be run the same way as always.

Change-Id: Id1e8233c5e366e6ec864fb435f7f3cbf290ffae0