Unsteady state in volume list should prevent cache
Volumes that are in states that imply a state change should cause us to
not cache. That way we are never spinning on a cache knowing a new state
is impending. This is only the first half of the change that is needed.
We also need to invalidate the volume list cache if we change anything.

Note that the odd structure for the should_cache_fn is imposed on us by
dogpile.cache's decorator function, which requires that things be fully
defined at parse time.

Change-Id: I8a34e3d04edea6c1b8ed636609102d7d0b7c86d8