Deprecate use of cache in list_volumes
The more proactive approach is to invalidate rather than use
cache=False. In all cases, we are about to update, or have just updated,
the item that is in the cache. This will provide more liveness for other
cache users with a shared cache, and will provide less skew between our
own queries to the cache versus reality.

Change-Id: I5a41a3d21fd451f5432daa4e268b57328a9ad90f