Fix major update_user issues
update_user didn't actually work at all, and needed porting to the
tasks system. Also it needs to invalidate the user cache to make sure
the latest list is requested.

Note that this breaks backward compatibility because defaulting the
enabled argument to True would mean we are unable to tell if the user
passed True, or left it blank. But the old method was entirely broken
anyway so we don't care at all.

Change-Id: Ief38f75b9b715cf149ada599c28cea03f99c8174