Move wait_for_status to resource module
wait_for_status is a function that is useful to many resources such as
volumes, stacks, databases, servers, and images. It would useful to have a
generic utility function to wait on the status of any such resource.

In the resource proxies the wait_for functions are named specifically to
that resource (e.g. compute.v2._proxy.wait_for_server) to stay consistent
with the verb_resource naming convention.

Change-Id: I6290235a9f7b0c204fc3933bf35497d0d8c4d0fe