Add early service fail and active check method
To make things clearer - do an active check at each client instantiation
to make sure that the service exists in the catalog so that we can throw
a clear error. Also, add a method that can be used for switching on
things like "if neutron: neutron-floating-ips: else: nova-floating-ips"
that doesn't have to do a REST roundtrip to find the failure.

Change-Id: I43b32fff56e09f31bd6d6845b3bf9b5414632651