Common list method for proxies
This change introduces a common list method for proxies, enabling the
use of query parameters to be used in limiting the data returned by a
request. In accepting query parameters, it also handles accepting
parameters that are resources by converting the resource into the
appropriate ID, so you can, e.g., get all servers that are running with
a particular image via that image's resource instance.

The Proxy._list method accept query parameters as keyword arguments that
are passed on to Resource.list, which are then passed on to
Resource.page, which are then passed on to session.get via the params=
parameters, which then causes the requests library to append those
key/value pairs onto the request URL as query parameters. From each
proxy's use of this common class, the approach was taken to accept the
query parameters in, e.g., compute.servers, as keyword arguments rather
than explicit parameters. This was done as there are a large amount of
parameters accepted by not just the /servers resource, but by many
others around other services.

As with other common methods for proxies, this change applies the base
method to the servers method (formerly known as list_servers).
Additionally, this method accepting query parameters to limit responses
will work in conjunction with the `get` methods with the aim of
converging the get/list/find set of methods to just get/list.

As with other changes like this, it introduces a second test
verification method that duplicates an existing one since we'll have two
APIs temporarily, but that will be removed once we've converged on this
API for listing.

Change-Id: I5e6d62a649a9f492433ea76d6eeaa754884e40ec